requirements.txt
alembic==1.1.0
blinker==1.4
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-http-client==3.3.1
requests==2.25.0
six==1.12.0
SQLAlchemy==1.3.8
starkbank-ecdsa==1.1.0
urllib3==1.26.2
visitor==0.1.3
Werkzeug==0.15.6
WTForms==2.2.1
